pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digital-archive-maker"
version = "0.9.0"
# NOTE: Bump to 1.0.0 for public release
description = "Physical media to digital library automation - CD/DVD/Blu-ray ripping, tagging, and sync"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Martin Diekhoff"}
]
keywords = [
    "media",
    "cd-ripping",
    "dvd",
    "blu-ray",
    "flac",
    "musicbrainz",
    "jellyfin",
    "plex",
    "metadata",
    "tagging"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "mutagen>=1.45.0",
    "musicbrainzngs>=0.7.1",
    "rapidfuzz>=2.0.0",
    "pyacoustid>=1.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "tqdm>=4.66.0",
    "lyricsgenius>=3.0.1",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.scripts]
dam = "dam.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/geniusworks/digital-archive-maker"
Documentation = "https://github.com/geniusworks/digital-archive-maker#readme"
Repository = "https://github.com/geniusworks/digital-archive-maker"
Issues = "https://github.com/geniusworks/digital-archive-maker/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dam*", "bin*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    _archive
    | _install
    | venv
    | \.git
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["_archive/*", "_install/*", "venv/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["_archive", "_install", "venv"]
